// src/server/gateway/src/proto/auth.pb.ts
export interface FieldError {
  property: string;
  message: string;
}

export interface CommonRes {
  code: number;
  status: string;
  message: FieldError[];
  data: { [key: string]: any } | undefined;
}

/** Register */
export interface RegisterReq {
  firstname: string;
  lastname: string;
  username: string;
  password: string;
  email: string;
}

export interface ResendOTPVerifyRegisterAccountReq {
  idRegister: number;
  email: string;
}

/** Verify OTP when register account */
export interface OTPVerifyRegisterAccountReq {
  email: string;
  otp: string;
}

/** Log in */
export interface LogInReq {
  username: string;
  password: string;
  latitude: number;
  longitude: number;
}

export interface LogOutReq {
  refreshToken: string;
}

/** Validate token */
export interface ValidateTokenReq {
  token: string;
  type: string;
}

/** Get user info */
export interface GetUserInfoReq {
  token: string;
  type: string;
}

/** Renew access token */
export interface RenewAccessTokenReq {
  token: string;
  type: string;
}

export const AUTH_PACKAGE_NAME = "auth";

wrappers[".google.protobuf.Struct"] = { fromObject: Struct.wrap, toObject: Struct.unwrap } as any;

export interface AuthServiceClient {
  register(request: RegisterReq): Observable<CommonRes>;

  resendOtpVerifyRegisterAccount(request: ResendOTPVerifyRegisterAccountReq): Observable<CommonRes>;

  otpVerifyRegisterAccount(request: OTPVerifyRegisterAccountReq): Observable<CommonRes>;

  logIn(request: LogInReq): Observable<CommonRes>;

  logOut(request: LogOutReq): Observable<CommonRes>;

  validateToken(request: ValidateTokenReq): Observable<CommonRes>;

  getUserInfo(request: GetUserInfoReq): Observable<CommonRes>;

  renewAccessToken(request: RenewAccessTokenReq): Observable<CommonRes>;
}
// ...
export const AUTH_SERVICE_NAME = "AuthService";